transitObjects.socketRoutes.ts: Extracted socket endpoint logic into a seperate file
