+++ packages/transition-backend/src/api/transitObjects.socketRoutes.ts
@@ -5,469 +5,121 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 import { EventEmitter } from 'events';
-import geobuf from 'geobuf';
-import Pbf from 'pbf';
+import { transitObjectEndpointDefinitions } from './definitions/transitObjects.definitions';
 
-import { isSocketIo } from './socketUtils';
 
-import AgencyCollection from 'transition-common/lib/services/agency/AgencyCollection';
-import LineCollection from 'transition-common/lib/services/line/LineCollection';
-import NodeCollection from 'transition-common/lib/services/nodes/NodeCollection';
-import PathCollection from 'transition-common/lib/services/path/PathCollection';
-import ScenarioCollection from 'transition-common/lib/services/scenario/ScenarioCollection';
-import ServiceCollection from 'transition-common/lib/services/service/ServiceCollection';
-import * as Status from 'chaire-lib-common/lib/utils/Status';
-
-import agenciesDbQueries from '../models/db/transitAgencies.db.queries';
-import linesDbQueries from '../models/db/transitLines.db.queries';
-import nodesDbQueries from '../models/db/transitNodes.db.queries';
-import pathsDbQueries from '../models/db/transitPaths.db.queries';
-import scenariosDbQueries from '../models/db/transitScenarios.db.queries';
-import servicesDbQueries from '../models/db/transitServices.db.queries';
-import schedulesDbQueries from '../models/db/transitSchedules.db.queries';
-
-import * as agenciesCacheQueries from '../models/capnpCache/transitAgencies.cache.queries';
-import * as linesCacheQueries from '../models/capnpCache/transitLines.cache.queries';
-import * as nodesCacheQueries from '../models/capnpCache/transitNodes.cache.queries';
-import * as pathsCacheQueries from '../models/capnpCache/transitPaths.cache.queries';
-import * as scenariosCacheQueries from '../models/capnpCache/transitScenarios.cache.queries';
-import * as servicesCacheQueries from '../models/capnpCache/transitServices.cache.queries';
-
-import TrError from 'chaire-lib-common/lib/utils/TrError';
-
-const transitClassesConfig = {
-    agencies: {
-        lowerCaseName: 'agency',
-        className: 'Agency',
-        classNamePlural: 'Agencies',
-        dbQueries: agenciesDbQueries,
-        cacheQueries: agenciesCacheQueries,
-        collection: new AgencyCollection([], {})
-    },
-    lines: {
-        lowerCaseName: 'line',
-        className: 'Line',
-        classNamePlural: 'Lines',
-        dbQueries: linesDbQueries,
-        cacheQueries: linesCacheQueries,
-        collection: new LineCollection([], {})
-    },
-    nodes: {
-        lowerCaseName: 'node',
-        className: 'Node',
-        classNamePlural: 'Nodes',
-        hasIntegerId: true,
-        dbQueries: nodesDbQueries,
-        cacheQueries: nodesCacheQueries,
-        collection: new NodeCollection([], {})
-    },
-    paths: {
-        lowerCaseName: 'path',
-        className: 'Path',
-        classNamePlural: 'Paths',
-        hasIntegerId: true,
-        dbQueries: pathsDbQueries,
-        cacheQueries: pathsCacheQueries,
-        collection: new PathCollection([], {})
-    },
-    scenarios: {
-        lowerCaseName: 'scenario',
-        className: 'Scenario',
-        classNamePlural: 'Scenarios',
-        dbQueries: scenariosDbQueries,
-        cacheQueries: scenariosCacheQueries,
-        collection: new ScenarioCollection([], {})
-    },
-    services: {
-        lowerCaseName: 'service',
-        className: 'Service',
-        classNamePlural: 'Services',
-        dbQueries: servicesDbQueries,
-        cacheQueries: servicesCacheQueries,
-        collection: new ServiceCollection([], {})
-    },
-    schedules: {
-        lowerCaseName: 'schedule',
-        className: 'Schedule',
-        classNamePlural: 'Schedules',
-        hasIntegerId: false,
-        dbQueries: schedulesDbQueries,
-        cacheQueries: {}
-    }
-};
-
-// TODO Add unit tests and typings when db queries and cache queries are refactored again. See if cache/db relation needs to be revisited now
-export const setupObjectSocketRoutes = (
-    socket: EventEmitter,
-    transitClassesConfig: {
-        [key: string]: {
-            lowerCaseName: string;
-            className: string;
-            classNamePlural: string;
-            hasIntegerId?: boolean;
-            dbQueries: any;
-            cacheQueries: any;
-            // TODO Used to temporarily load collections. needed?
-            collection?: any;
-        };
-    }
-) => {
-    for (const lowerCasePlural in transitClassesConfig) {
-        const transitClassConfig = transitClassesConfig[lowerCasePlural];
+function setupObjectSocketRoutes(socket: EventEmitter) {
+    for (const lowerCasePlural in transitObjectEndpointDefinitions) {
+        const endpointDefinitions = transitObjectEndpointDefinitions[lowerCasePlural]
 
         // Create a new object
-        socket.on(`transit${transitClassConfig.className}.create`, async (attributes, callback) => {
-            try {
-                const returningArray = transitClassConfig.hasIntegerId ? ['id', 'integer_id'] : ['id'];
-                const returning = await transitClassConfig.dbQueries.create(attributes, returningArray);
-                if (transitClassConfig.hasIntegerId) {
-                    attributes.integer_id = returning.integer_id;
-                }
-                if (isSocketIo(socket)) {
-                    socket.broadcast.emit('data.updated');
-                }
-                if (transitClassConfig.cacheQueries.objectToCache) {
-                    try {
-                        await transitClassConfig.cacheQueries.objectToCache(
-                            attributes,
-                            attributes.data.customCachePath
-                        );
-                        callback({
-                            ...returning
-                        });
-                    } catch (error) {
-                        throw new TrError(
-                            `cannot save cache file ${transitClassConfig.className} because of an error: ${error}`,
-                            'SKTTRRD0001',
-                            'CacheCouldNotBeSavedBecauseError'
-                        );
-                    }
-                } else {
-                    socket.emit('cache.dirty');
-                    callback({
-                        ...returning
-                    });
-                }
-            } catch (error) {
-                console.error(error);
-                callback(TrError.isTrError(error) ? error.export() : { error });
-            }
+        socket.on(`transit${endpointDefinitions.className}.create`, async (attributes, callback) => {
+            const response = await endpointDefinitions.create(socket, attributes);
+            callback(response);
         });
 
         // Read the object from the database
         socket.on(
-            `transit${transitClassConfig.className}.read`,
+            `transit${endpointDefinitions.className}.read`,
             async (id: string, customCachePath: string | undefined, callback) => {
-                try {
-                    const object = await transitClassConfig.dbQueries.read(id);
-                    callback({
-                        [transitClassConfig.lowerCaseName]: object.attributes ? object.attributes : object
-                    });
-                } catch (error) {
-                    console.error(error);
-                    callback(TrError.isTrError(error) ? error.export() : { error });
-                }
+                const response = await endpointDefinitions.read(id, customCachePath);
+                callback(response);
             }
         );
 
         // Update the object in the database and cache if required
-        socket.on(`transit${transitClassConfig.className}.update`, async (id: string, attributes, callback) => {
-            try {
-                const updatedId = await transitClassConfig.dbQueries.update(id, attributes);
-                if (isSocketIo(socket)) {
-                    socket.broadcast.emit('data.updated');
-                }
-                if (transitClassConfig.cacheQueries.objectToCache) {
-                    try {
-                        await transitClassConfig.cacheQueries.objectToCache(
-                            attributes,
-                            attributes.data.customCachePath
-                        );
-                        callback({
-                            id: updatedId
-                        });
-                    } catch (error) {
-                        throw new TrError(
-                            `cannot update cache file ${transitClassConfig.className} because of an error: ${error}`,
-                            'SKTTRUP0001',
-                            'CacheCouldNotBeUpdatedBecauseError'
-                        );
-                    }
-                } else {
-                    socket.emit('cache.dirty');
-                    callback({
-                        id: updatedId
-                    });
-                }
-            } catch (error) {
-                console.error(error);
-                callback(TrError.isTrError(error) ? error.export() : { error });
-            }
+        socket.on(`transit${endpointDefinitions.className}.update`, async (id: string, attributes, callback) => {
+            const response = await endpointDefinitions.update(socket, id, attributes);
+            callback(response);
         });
 
         // Delete the object from database and cache if required
         socket.on(
-            `transit${transitClassConfig.className}.delete`,
+            `transit${endpointDefinitions.className}.delete`,
             async (id: string, customCachePath: string | undefined, callback) => {
-                try {
-                    const deletedId = await transitClassConfig.dbQueries.delete(id);
-                    if (deletedId === undefined) {
-                        // The object was not deleted
-                        callback(Status.createOk({ id: undefined }));
-                        return;
-                    }
-                    if (isSocketIo(socket)) {
-                        socket.broadcast.emit('data.updated');
-                    }
-                    if (transitClassConfig.cacheQueries.deleteObjectCache) {
-                        try {
-                            await transitClassConfig.cacheQueries.deleteObjectCache(id, customCachePath);
-                            callback(Status.createOk({ id: deletedId }));
-                        } catch (error) {
-                            throw new TrError(
-                                `cannot delete cache file ${transitClassConfig.className} because of an error: ${error}`,
-                                'SKTTRDL0001',
-                                'CacheCouldNotBeDeletedBecauseError'
-                            );
-                        }
-                    } else {
-                        socket.emit('cache.dirty');
-                        callback(Status.createOk({ id: deletedId }));
-                    }
-                } catch (error) {
-                    console.error(error);
-                    callback(Status.createError(TrError.isTrError(error) ? error.message : 'Error deleting object'));
-                }
+                const response = await endpointDefinitions.delete(socket, id, customCachePath);
+                callback(response);
             }
         );
 
         // Get the geojson collection from DB if there is a geojson collection function
-        if (transitClassConfig.dbQueries.geojsonCollection) {
+        if (endpointDefinitions.geojsonCollection) {
             socket.on(
-                `transit${transitClassConfig.classNamePlural}.geojsonCollection`,
+                `transit${endpointDefinitions.classNamePlural}.geojsonCollection`,
                 async (params = { format: 'geojson' }, callback) => {
-                    try {
-                        const geojson = await transitClassConfig.dbQueries.geojsonCollection(params);
-                        if (params.format === 'geobuf') {
-                            const geobufjson = Buffer.from(geobuf.encode(geojson, new Pbf()));
-                            callback({ geobuf: geobufjson });
-                        } else {
-                            callback({ geojson });
-                        }
-                    } catch (error) {
-                        console.error(error);
-                        callback(TrError.isTrError(error) ? error.export() : { error });
-                    }
+                    const response = await endpointDefinitions.geojsonCollection!(params);
+                    callback(response);
                 }
             );
         }
 
         // Get the collection from DB if there is a collection function
-        if (transitClassConfig.dbQueries.collection) {
-            socket.on(`transit${transitClassConfig.classNamePlural}.collection`, async (dataSourceId, callback) => {
-                try {
-                    const collection = await transitClassConfig.dbQueries.collection();
-                    callback({ collection });
-                } catch (error) {
-                    console.error(error);
-                    callback(TrError.isTrError(error) ? error.export() : { error });
-                }
+        if (endpointDefinitions.collection) {
+            socket.on(`transit${endpointDefinitions.classNamePlural}.collection`, async (dataSourceId, callback) => {
+                const response = await endpointDefinitions.collection!(dataSourceId);
+                callback(response);
             });
         }
 
         // Save an object to cache
         // TODO Saving an object to cache is included in the create and update routes. And now there is not much that is not in the database (transferable nodes for instance), this route could be removed
-        if (transitClassConfig.cacheQueries.objectToCache) {
-            socket.on(`transit${transitClassConfig.className}.saveCache`, async (attributes, callback) => {
-                try {
-                    await transitClassConfig.cacheQueries.objectToCache(attributes, attributes.data.customCachePath);
-                    callback({});
-                } catch (error) {
-                    console.error(error);
-                    callback(
-                        new TrError(
-                            `cannot save cache file ${transitClassConfig.className} because of an error: ${error}`,
-                            'SKTTRSVC0001',
-                            'CacheCouldNotBeSavedBecauseError'
-                        ).export()
-                    );
-                }
+        if (endpointDefinitions.saveCache) {
+            socket.on(`transit${endpointDefinitions.className}.saveCache`, async (attributes, callback) => {
+                const response = await endpointDefinitions.saveCache!(attributes);
+                callback(response);
             });
         }
 
         // Delete object from cache if required
         // TODO Included in the call to delete, the individual call should not exist
-        if (transitClassConfig.cacheQueries.deleteObjectCache) {
+        if (endpointDefinitions.deleteCache) {
             socket.on(
-                `transit${transitClassConfig.className}.deleteCache`,
+                `transit${endpointDefinitions.className}.deleteCache`,
                 async (id: string, customCachePath: string | undefined, callback) => {
-                    try {
-                        await transitClassConfig.cacheQueries.deleteObjectCache(id, customCachePath);
-                        callback({});
-                    } catch (error) {
-                        console.error(error);
-                        callback(
-                            new TrError(
-                                `cannot delete cache file ${transitClassConfig.className} because of an error: ${error}`,
-                                'SKTTRDLC0001',
-                                'CacheCouldNotBeDeletedBecauseError'
-                            ).export()
-                        );
-                    }
+                    const response = await endpointDefinitions.deleteCache!(id, customCachePath);
+                    callback(response);
                 }
             );
         }
 
-        if (transitClassConfig.cacheQueries.deleteObjectsCache) {
+        if (endpointDefinitions.deleteMultipleCache) {
             socket.on(
-                `transit${transitClassConfig.className}.deleteMultipleCache`,
-                (ids, customCachePath, callback) => {
-                    transitClassConfig.cacheQueries
-                        .deleteObjectsCache(ids, customCachePath)
-                        .then(() => {
-                            callback({
-                                error: null
-                            });
-                        })
-                        .catch((error) => {
-                            console.error(error);
-                            callback(
-                                new TrError(
-                                    `cannot delete cache files ${transitClassConfig.className} because of an error: ${error}`,
-                                    'SKTTRDLC0002',
-                                    'CacheCouldNotBeDeletedBecauseError'
-                                ).export()
-                            );
-                        });
+                `transit${endpointDefinitions.className}.deleteMultipleCache`,
+                async (ids, customCachePath, callback) => {
+                    const response = await endpointDefinitions.deleteMultipleCache!(ids, customCachePath);
+                    callback(response);
                 }
             );
         }
 
         // Load an object from the cache if available
-        if (transitClassConfig.cacheQueries.objectFromCache) {
+        if (endpointDefinitions.loadCache) {
             socket.on(
-                `transit${transitClassConfig.className}.loadCache`,
+                `transit${endpointDefinitions.className}.loadCache`,
                 async (id: string, customCachePath: string | undefined, callback) => {
-                    try {
-                        const object = await transitClassConfig.cacheQueries.objectFromCache(id, customCachePath);
-                        callback({
-                            [transitClassConfig.lowerCaseName]: object && object.attributes ? object.attributes : object
-                        });
-                    } catch (error) {
-                        console.error(error);
-                        callback(
-                            new TrError(
-                                `cannot load cache file ${transitClassConfig.className} because of an error: ${error}`,
-                                'SKTTRLDC0001',
-                                'CacheCouldNotBeLoadedBecauseError'
-                            ).export()
-                        );
-                    }
+                    const response = await endpointDefinitions.loadCache!(id, customCachePath);
+                    callback(response);
                 }
             );
         }
 
-        if (transitClassConfig.cacheQueries.collectionToCache) {
+        if (endpointDefinitions.saveCollectionCache) {
             socket.on(
-                `transit${transitClassConfig.classNamePlural}.saveCollectionCache`,
-                (collection = null, customCachePath, callback) => {
-                    if (collection) {
-                        transitClassConfig.cacheQueries
-                            .collectionToCache(collection, customCachePath)
-                            .then(() => {
-                                callback({
-                                    error: null
-                                });
-                            })
-                            .catch((error) => {
-                                throw new TrError(
-                                    `cannot save cache collection file ${transitClassConfig.classNamePlural} because of an error: ${error}`,
-                                    'SKTTRSGC0001',
-                                    'CacheCouldNotBeSavedBecauseError'
-                                );
-                            });
-                    } else if (transitClassConfig.dbQueries.geojsonCollection) {
-                        transitClassConfig.dbQueries
-                            .geojsonCollection()
-                            .then((collection) => {
-                                transitClassConfig.collection.loadFromCollection(collection.features);
-                                transitClassConfig.cacheQueries
-                                    .collectionToCache(transitClassConfig.collection, customCachePath)
-                                    .then(() => {
-                                        callback({
-                                            error: null
-                                        });
-                                    })
-                                    .catch((error) => {
-                                        throw new TrError(
-                                            `cannot save cache collection file ${transitClassConfig.classNamePlural} because of an error: ${error}`,
-                                            'SKTTRSGC0002',
-                                            'CacheCouldNotBeSavedBecauseError'
-                                        );
-                                    });
-                            })
-                            .catch((error) => {
-                                throw new TrError(
-                                    `cannot save cache collection file ${transitClassConfig.classNamePlural} because of an error: ${error}`,
-                                    'SKTTRSGC0003',
-                                    'CacheCouldNotBeSavedBecauseError'
-                                );
-                            });
-                    } else {
-                        transitClassConfig.dbQueries
-                            .collection()
-                            .then((collection) => {
-                                transitClassConfig.collection.loadFromCollection(collection);
-                                transitClassConfig.cacheQueries
-                                    .collectionToCache(transitClassConfig.collection, customCachePath)
-                                    .then(() => {
-                                        callback({
-                                            error: null
-                                        });
-                                    })
-                                    .catch((error) => {
-                                        throw new TrError(
-                                            `cannot save cache collection file ${transitClassConfig.classNamePlural} because of an error: ${error}`,
-                                            'SKTTRSC0001',
-                                            'CacheCouldNotBeSavedBecauseError'
-                                        );
-                                    });
-                            })
-                            .catch((error) => {
-                                throw new TrError(
-                                    `cannot save cache collection file ${transitClassConfig.classNamePlural} because of an error: ${error}`,
-                                    'SKTTRSC0002',
-                                    'CacheCouldNotBeSavedBecauseError'
-                                );
-                            });
-                    }
+                `transit${endpointDefinitions.classNamePlural}.saveCollectionCache`,
+                async (collection = null, customCachePath, callback) => {
+                    const response = await endpointDefinitions.saveCollectionCache!(collection, customCachePath);
+                    callback(response);
                 }
             );
         }
 
         // TODO Do we still need to load an entire collection from cache. Cache should be one-way?
-        if (transitClassConfig.cacheQueries.collectionFromCache) {
+        if (endpointDefinitions.loadCollectionCache) {
             socket.on(
-                `transit${transitClassConfig.classNamePlural}.loadCollectionCache`,
-                (customCachePath, callback) => {
-                    transitClassConfig.cacheQueries
-                        .collectionFromCache(customCachePath)
-                        .then((collection) => {
-                            callback({
-                                collection
-                            });
-                        })
-                        .catch((error) => {
-                            if (!error.export) {
-                                throw new TrError(
-                                    `cannot load cache collection file ${transitClassConfig.classNamePlural} because of an error: ${error}`,
-                                    'SKTTRLC0001',
-                                    'CacheCouldNotBeLoadedBecauseError'
-                                );
-                            }
-                            console.error(error);
-                            callback(error.export());
-                        });
+                `transit${endpointDefinitions.classNamePlural}.loadCollectionCache`,
+                async (customCachePath, callback) => {
+                    const response = await endpointDefinitions.loadCollectionCache!(customCachePath);
+                    callback(response);
                 }
             );
         }
@@ -476,5 +128,5 @@ export const setupObjectSocketRoutes = (
 
 // Add operations on object socket routes for each object of Transition
 export default function (socket: EventEmitter) {
-    setupObjectSocketRoutes(socket, transitClassesConfig);
+    setupObjectSocketRoutes(socket);
 }
